2.26.2 to 2.27.1 source update

+++ mantisbt/core/access_api.php
@@ -213,7 +213,7 @@ function access_get_global_level( $p_user_id = null ) {
 		return false;
 	}
 
-	return user_get_field( $p_user_id, 'access_level' );
+	return (int)user_get_field( $p_user_id, 'access_level' );
 }
 
 /**
@@ -354,7 +354,6 @@ function access_has_project_level( $p_access_level, $p_project_id = null, $p_use
  * @param integer               $p_limit        Maximum number of results, default is 0 for all results
  *
  * @return array                The filtered array of project ids
- * @throws ClientException
  */
 function access_project_array_filter( $p_access_level, array $p_project_ids = null, $p_user_id = null, $p_limit = 0 ) {
 	# Short circuit the check in this case
